poc1/did_kyc/requirements.txt
Flask == 1.1.2
PyNaCl == 1.4.0
configparser == 5.0.0
Flask-Cors==3.0.9